Game almost complete... Need to fix some issues with the flyweight pattern... Specs not yet written...(BAD)

+++ MineSweeper_Storyboards/Common.h
@@ -11,5 +11,5 @@
 #import "EmptySpot.h"
 #import "MineSpot.h"
 
-static const int cols = 8;
-static const int rows = 7;
+static const NSInteger cols = 8;
+static const NSInteger rows = 7;
